backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
qdrant-client>=1.12.0
fastembed>=0.3.0
python-frontmatter>=1.1.0
tiktoken>=0.8.0
pydantic>=2.9.0
python-dotenv>=1.0.0
slowapi>=0.1.9
httpx>=0.27.0
rank-bm25>=0.2.2
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
